New script rr_tune_survey.py that runs the Recycler ring.
A single bunch of
41 particles is launched at dp/p offsets corresponding to a frequency shift of
arange(-2000, 2050, 100) Hz.

The turn-by-turn coordinates of each particle is saved in the HDF5 file tracks.h5
in dataset named track_coords. This is an array of shape (nturns, nparticles, 7).
The run is currently set for 2000 turns so nturns = 2001 (initial + 2000).
nparticles is 41 for a particle at each of the different momenta, and the third
dimension are the phase space coordinates of the particle
(x, x', y, y', cdt, dp/p, label).

This can be easily read with the h5py python module.